syntax/attr: reduce reliance on parser
