+++ src/libsyntax/lib.rs
@@ -96,8 +96,7 @@ pub mod json;
 pub mod ast;
 pub mod attr;
 pub mod source_map;
-#[macro_use]
-pub mod config;
+#[macro_use] pub mod config;
 pub mod entry;
 pub mod feature_gate;
 pub mod mut_visit;
